# This is a combination of 13 commits. # The first commit's message is: added date class
updates to add more of the fields in.

Conflicts:
	app/assets/javascripts/application.js
	app/assets/javascripts/components.js
	app/assets/javascripts/components/OpenSeadragonViewer.jsx
	app/assets/javascripts/components/ReactDropzone.jsx
	app/assets/javascripts/components/Thumbnail.jsx
	app/assets/javascripts/components/showcase_editor/ShowcaseEditor.jsx
	package.json

sorted out the flow control for the component.

make the empty text better.

add the date saving

add the other optional date fields

added padding to the inline form.

# This is the 2nd commit message:

change the way the component is tagged to be more correct

# This is the 3rd commit message:

update to display the new fields in the meta data hash in the item json.

# This is the 4th commit message:

added a specific date class to do the parsing of the metadata dates.

# This is the 5th commit message:

update date formatter to display the correct format.

# This is the 6th commit message:

updated the metadata object to use the metadata date.

# This is the 7th commit message:

add the date display code to the api.

# This is the 8th commit message:

DEC-397 Add date validation for date metadata fields

# This is the 9th commit message:

Hound

# This is the 10th commit message:

Hound

# This is the 11th commit message:

fix schema for this branch

# This is the 12th commit message:

fix suggestions

# This is the 13th commit message:

rename for consistency